% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blossom.R
\name{blossom}
\alias{blossom}
\title{Blossom's algorithm}
\usage{
blossom(G, weighted = FALSE, maxcardinality = FALSE)
}
\arguments{
\item{G}{The graph to compute the maximum cardinality matching}

\item{weighted}{Whether the graph is weighted, if true, weights should be obtained by E(G)$weight}

\item{maxcardinality}{Whether the maximum weight should be computed over all maximum cardinality matchings}
}
\value{
The maximum weighted matching for G, in a list of edges
}
\description{
Computes the weighted bipartite matching using Hungarian's algorithm
}
\details{
Blossom's algorithm for maximum cardinality matching for general graph

Efficiently compute the maximum weighted biparitite matching using the Hungarian algorithm (TODO: citation)
Almost a direct port of Joris van Rantwijk's python code at http://jorisvr.nl/files/graphmatching/20130407/mwmatching.py
}

